requirements.txt
requests~=2.31.0
Flask~=3.0.2
python-docx~=1.1.0
pandas~=2.2.0
pillow~=10.2.0

